sndjvu_codec: fix compilation
This uses a trick with ManuallyDrop that steffahn came up with on URLO.
Ideally it wouldn't be needed, but the compiler seems to have some blind
spots here. I do plan to look for other workarounds/restructurings that
might make it easier for sndjvu_codec consumers to write custom BZZ
compression loops.